edit

+++ 0x15-file_io/1-create_file.c
@@ -9,7 +9,7 @@ int create_file(const char *filename, char *text_content)
 {
 	int fd;
 	int i;
-	ssize_t _bytes = 0;
+	int _bytes = 0;
 
 	if (filename == NULL)
 		return (-1);
